requirements.txt
bcrypt==4.0.1
blinker==1.7.0
cffi==1.16.0
click==8.1.7
colorama==0.4.6
cryptography==41.0.5
dnspython==2.4.2
email-validator==2.1.0.post1
et-xmlfile==1.1.0
Flask==3.0.0
Flask-Admin==1.6.1
Flask-Login==0.6.3
flask-mailman==1.0.0
Flask-Principal==0.4.0
Flask-Security-Too==5.3.2
Flask-SQLAlchemy==3.1.1
Flask-WTF==1.2.1
greenlet==3.0.1
idna==3.4
importlib-metadata==6.8.0
importlib-resources==6.1.1
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.3
numpy==1.26.2
openpyxl==3.1.2
pandas==2.1.3
passlib==1.7.4
Pillow==10.1.0
pycparser==2.21
PyMySQL==1.1.0
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3.post1
six==1.16.0
SQLAlchemy==2.0.23
typing_extensions==4.8.0
tzdata==2023.3
Werkzeug==3.0.1
WTForms==3.1.1
zipp==3.17.0